+++ lib/airbrake/rails.rb
@@ -7,6 +7,8 @@ module Airbrake
       initializer('airbrake.middleware') do |app|
         require 'airbrake/rails/action_controller_route_subscriber'
         require 'airbrake/rails/action_controller_notify_subscriber'
+        require 'airbrake/rails/action_controller_performance_breakdown_subscriber'
+
         require 'airbrake/rails/active_record_subscriber' if defined?(ActiveRecord)
 
         # Since Rails 3.2 the ActionDispatch::DebugExceptions middleware is